Optimize fontification by avoiding recalculating a constant regexp
regexp-opt can be slow because it also tries to optimize the
regexp. For example:

    (regexp-opt (list "aaa" "abb"))
    ;; -> "\\(?:a\\(?:aa\\|bb\\)\\)"

So it shouldn't be run every time a fontification happens.

This does mean that when
lisp-extra-font-lock-loop-keywords-with-var and
lisp-extra-font-lock-loop-keywords are changed, the changes would
only take effect after the file is reloaded. This shouldn't be a
problem though, as these variables are not meant to be changed, and
are unlikely to be changed anyways.